# api-laravel-main/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IFRI-UAC notation, API Documentation",
        "description": "IFRI-UAC notation, API Documentation",
        "contact": {
            "email": "contact@equipe-api.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.43.5:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/annees-academiques": {
            "get": {
                "tags": [
                    "AnneeAcademique"
                ],
                "summary": "Années Académiques - List",
                "description": "Returns list années académqiues",
                "operationId": "anneeAcademiqueList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AnneeAcademiqueModel"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/annees-academiques/create": {
            "post": {
                "tags": [
                    "AnneeAcademique"
                ],
                "summary": "AnneeAcademique - Create",
                "description": "Returns AnneeAcademique created",
                "operationId": "createAnneeAcademique",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnneeAcademiqueRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnneeAcademiqueResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/annees-academiques/{id}/update": {
            "put": {
                "tags": [
                    "AnneeAcademique"
                ],
                "summary": "AnneeAcademique - Update",
                "description": "Returns updated AnneeAcademique",
                "operationId": "updateAnneeAcademique",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "AnneeAcademique id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnneeAcademiqueRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnneeAcademiqueResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/annees-academiques/{id}/delete": {
            "delete": {
                "tags": [
                    "AnneeAcademique"
                ],
                "summary": "AnneeAcademique - Delete",
                "description": "Returns null AnneeAcademique",
                "operationId": "deleteAnneeAcademique",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "AnneeAcademique id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnneeAcademiqueResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/users/generate-credentials": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authentification - Generate credentials",
                "description": "Returns users list",
                "operationId": "generateCredentials",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CredentialRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CredentialModel"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AnneeAcademiqueModel": {
                "title": "AnneeAcademique Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identifier",
                        "type": "integer",
                        "format": "int64",
                        "example": 18
                    },
                    "libelle": {
                        "title": "libelle",
                        "description": "libelle of AnneeAcademique",
                        "type": "string",
                        "example": "2020-2021"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AnneeAcademique"
                }
            },
            "AnneeAcademiqueRequest": {
                "title": "Save AnneeAcademique Request",
                "required": [
                    "libelle"
                ],
                "properties": {
                    "libelle": {
                        "title": "libelle",
                        "description": "Libelle of the AnneeAcademique",
                        "type": "string",
                        "example": "2020-2021"
                    }
                },
                "type": "object"
            },
            "AnneeAcademiqueResponse": {
                "title": "AnneeAcademique Response",
                "properties": {
                    "anneeAcademique": {
                        "$ref": "#/components/schemas/AnneeAcademiqueModel"
                    },
                    "message": {
                        "title": "message",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AnneeAcademiqueResponse"
                }
            },
...
            "CredentialModel": {
                "title": "Credential Model",
                "properties": {
                    "qrcode": {
                        "title": "qrcode",
                        "description": "CodeQr value",
                        "type": "string",
                        "example": "Koakdkvardnh19373JDJSJK72736YYSHJHJHDUY3676UJHD267ESJS"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Credential"
                }
            },
            "CredentialRequest": {
                "title": "Generate credentials Request",
                "required": [
                    "classe_id",
                    "effectif"
                ],
                "properties": {
                    "classe_id": {
                        "title": "classe_id",
                        "description": "Classe id",
                        "type": "integer",
                        "example": "2"
                    },
                    "effectif": {
                        "title": "effectif",
                        "description": "Effectif of the classe",
                        "type": "integer",
                        "example": "57"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "The value of access_token received from login: Ex: Bearer access_token_value",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}